Added test scenarios showing that there is a basic issue with how the analyzer is bootstrapped
The fact that our analyzer cannot infer purity for `DateTimeImmutable#getTimestamp()`, or that
it cannot find the `JSON_THROW_ON_ERROR` constant, indicates that there is something with how we
set up the static analysis toolchain, or how we bootstrap it.

For now, these tests mostly highlight the issue, while we wait for the upstream discussion
at https://github.com/vimeo/psalm/issues/5764#issuecomment-841174672 to come with some further
steps that we could potentially follow.