Reproducing issue in which, after the first mutant is scanned, further scans mutants with references to core PHP symbols fail
The reason for this failure seems to be corrupt state within `vimeo/psalm`'s `\Psalm\Internal\Codebase\Methods#$classlikes`
storage, which contains information about `DateTimeImmutable`, but incomplete, and only available when performing a fresh
scan from scratch.

The internal psalm in-memory storage gets corrupted while reading the first mutant (no references to PHP symbols), where
perhaps Psalm fails to load symbols that aren't necessary, and then the storage is never touched again ("warmed up"),
and therefore is a snapshot of an incomplete API for the referenced symbols.

That means that the second mutant being inspected by psalm references methods that are not declared, according to psalm
(actual bug).

Ref: https://github.com/vimeo/psalm/issues/5764#issuecomment-843423337
Ref: https://github.com/vimeo/psalm/issues/5764#issuecomment-843436051